requirements.txt
django
django-allauth
beautifulsoup4
